requirements.txt
sentence-transformers>=2.2.0
chromadb>=0.4.0
tiktoken>=0.5.0
numpy>2.0
google-generativeai>=0.8.6
python-dotenv
yt-dlp
assemblyai
ffmpeg
